% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{convertBase}
\alias{convertBase}
\title{Title Convertir base de datos}
\usage{
convertBase(file = file, sp, tipo, nameFileOut, ...)
}
\arguments{
\item{sp}{nombre de la especie}

\item{tipo}{Tipo de flota o arte}

\item{nameFileOut}{Nombre con el cual el archivo será guardado}

\item{...}{Argumentos extras}
}
\value{
Base de datos con el formato de captura y esfuerzo
}
\description{
Title Convertir base de datos
}

