% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(file, type, toTons, ...)
}
\arguments{
\item{file}{Dirección del archivo con información de capturas.}

\item{type}{Definir si se desea obtener estadísticos de Desembarques (\code{landings}),
Esfuerzo (\code{effort}) o CPUE (\code{cpue}).}

\item{toTons}{¿Desea tranformar los valores a toneladas? (\code{TRUE} or \code{FALSE}).}

\item{...}{Argumentos extras.}
}
\description{
getData
}
\details{
Esta función retornará un objeto de clase \code{landings}, \code{effort} o
\code{cpue}, dependiendo de lo seleccionado en el argumento \code{type}. La manera
cómo se generarán las figuras dependerá de cada clase.
}
\examples{
\dontrun{
getData(file = "file.csv", type = "effort")
}
}

