% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{getTable}
\alias{getTable}
\title{Función para obtener tablas resumen}
\usage{
getTable(data)
}
\arguments{
\item{data}{Objeto de clase \code{landings}, \code{effort} o \code{cpue} desde donde
se desea obtener la tabla resumen.}
}
\description{
Función para obtener tablas resumen
}

