% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{getTable2}
\alias{getTable2}
\title{Función para obtener tablas resumen mostrando las estadisticas quincenales dle último mes}
\usage{
getTable2(data)
}
\arguments{
\item{Objeto}{de clase \code{landings} o \code{effort} desde donde
se desea obtener la tabla resumen.}
}
\description{
Función para obtener tablas resumen mostrando las estadisticas quincenales dle último mes
}

