% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-class.R
\name{plot.cpue}
\alias{plot.cpue}
\title{Método \code{plot} para objetos de clase \code{cpue}.}
\usage{
\method{plot}{cpue}(x, time = NULL, ...)
}
\arguments{
\item{x}{Objeto de clase \code{cpue}.}

\item{time}{Argumento que hace referencia al paso de tiempo para ser graficados,
\code{"day"}, \code{"month"}, \code{"year"}.}

\item{...}{Argumentos extra.}
}

