% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort-class.R
\name{plot.effort}
\alias{plot.effort}
\title{Método \code{plot} para objetos de clase \code{effort}.}
\usage{
\method{plot}{effort}(x, time = NULL, ...)
}
\arguments{
\item{x}{Objeto de clase \code{effort}.}

\item{...}{Argumentos extra.}
}

