% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings-class.R
\name{print.landings}
\alias{print.landings}
\title{Método \code{print} para objetos de clase \code{landings}.}
\usage{
\method{print}{landings}(x, ...)
}
\arguments{
\item{...}{Argumentos extra.}

\item{object}{Objeto de clase \code{landings}.}
}

