% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-class.R
\name{print.summary.cpue}
\alias{print.summary.cpue}
\title{Método \code{print} para objetos de clase \code{summary.cpue}.}
\usage{
\method{print}{summary.cpue}(x, ...)
}
\arguments{
\item{x}{Objeto de clase \code{summary.cpue}.}

\item{...}{Argumentos extra.}
}

