% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort-class.R
\name{print.summary.effort}
\alias{print.summary.effort}
\title{Método \code{print} para objetos de clase \code{summary.effort}.}
\usage{
\method{print}{summary.effort}(x, ...)
}
\arguments{
\item{x}{Objeto de clase \code{summary.effort}.}

\item{...}{Argumentos extra.}
}

