% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings-class.R
\name{print.summary.landings}
\alias{print.summary.landings}
\title{Método \code{print} para objetos de clase \code{summary.landings}.}
\usage{
\method{print}{summary.landings}(x, ...)
}
\arguments{
\item{x}{Objeto de clase \code{summary.landings}.}

\item{...}{Argumentos extra.}
}

