% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{RepSeg-package}
\alias{RepSeg}
\alias{RepSeg-package}
\title{Herramientas para la generación de reportes de captura, esfuerzo y CPUE.}
\description{
Herramientas para la generación de reportes de captura, esfuerzo y CPUE.
}
\author{
Josymar Torrejón-Magallanes, \email{jotorrejon@imarpe.gob.pe}
}
\keyword{CPUE}
\keyword{captura,}
\keyword{esfuerzo,}
\keyword{reportes,}
\keyword{seguimiento,}

