% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-class.R
\name{summary.cpue}
\alias{summary.cpue}
\title{Método \code{summary} para objetos de clase \code{cpue}.}
\usage{
\method{summary}{cpue}(object, ...)
}
\arguments{
\item{object}{Objeto de clase \code{cpue}.}

\item{...}{Argumentos extra.}
}

