% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort-class.R
\name{summary.effort}
\alias{summary.effort}
\title{Método \code{summary} para objetos de clase \code{effort}.}
\usage{
\method{summary}{effort}(object, ...)
}
\arguments{
\item{object}{Objeto de clase \code{effort}.}

\item{...}{Argumentos extra.}
}

