% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings-class.R
\name{summary.landings}
\alias{summary.landings}
\title{Método \code{summary} para objetos de clase \code{landings}.}
\usage{
\method{summary}{landings}(object, ...)
}
\arguments{
\item{object}{Objeto de clase \code{landings}.}

\item{...}{Argumentos extra.}
}

