% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(file, type, toTons, ...)
}
\arguments{
\item{file}{Nombre del archivo coninformación de captura y cpue.}

\item{type}{Definir si se desea obtener estadísticos de Desembarques (\code{landings}),
Esfuerzo (\code{effort}) o CPUE (\code{cpue}).}

\item{toTons}{Para convertir los valores a toneladas (\code{TRUE} or \code{FALSE}).
Solo para desembarques (\code{landings}) y cpue (\code{cpue}).}

\item{...}{Argumentos extras.}
}
\description{
getData
}
\details{
Esta función retornará un objeto de clase \code{landings}, \code{effort} o
\code{cpue}, dependiendo de lo seleccionado en el argumento \code{type}. La manera
cómo se generarán las figuras dependerá de cada clase.
}
\examples{
\dontrun{
getData(file = "file.csv", type = "effort")
}
}

