% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{makeReport}
\alias{makeReport}
\title{Función que construye el reporte de seguiumiento}
\usage{
makeReport(desembarque, filename = NULL, openAtEnd = TRUE, paper = "a4r",
  width = 10, height = 10, ...)
}
\arguments{
\item{desembarque}{Objeto de clase \code{landings} de donde se obtendrán los datos de
desembarque.}

\item{filename}{Archivo en donde se guardará el pdf. Si es \code{NULL}, se creará un archivo temporal
que se mostrará al final.}

\item{openAtEnd}{¿Desea abrir el archivo al final del proceso? Si \code{filename = NULL}, este parámetro
no se tomará en cuenta.}

\item{...}{Argumentos extra pasados a la función \code{pdf}.}
}
\description{
Función que construye el reporte de seguiumiento
}
\examples{
\dontrun{
desembarque <- getData(file = "file.csv", type = "effort")
makeReport(desembarque)
}
}

