% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings-class.R
\name{plot.landings}
\alias{plot.landings}
\title{Método \code{plot} para objetos de clase \code{landings}.}
\usage{
\method{plot}{landings}(x, time = NULL, ...)
}
\arguments{
\item{x}{Objeto de clase \code{landings}.}

\item{time}{Argumento que hace referencia al paso de tiempo para ser graficados,
\code{"day"}, \code{"month"}, \code{"year"}.}

\item{...}{Argumentos extra.}
}

