% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort-class.R
\name{print.effort}
\alias{print.effort}
\title{Método \code{print} para objetos de clase \code{effort}.}
\usage{
\method{print}{effort}(x, ...)
}
\arguments{
\item{...}{Argumentos extra.}

\item{object}{Objeto de clase \code{effort}.}
}
\value{
Información básica en la base de datos: Número de registros, meses, años, número de puertos.
}

